reorganize files, created script to run everything
